create submodules; move data and tree into their own folders
