+++ docs/api/happler.rst
@@ -13,18 +13,42 @@ Command line interface
 Module contents
 ---------------
 
-happler.Data module
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+happler.data.data module
+~~~~~~~~~~~~~~~~~~~~~~~~
 
-.. automodule:: happler.Data
+.. automodule:: happler.data.data
    :members:
    :undoc-members:
    :show-inheritance:
 
-happler.Tree module
-~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+happler.data.genotypes module
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
-.. automodule:: happler.Tree
+.. automodule:: happler.data.genotypes
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+happler.data.phenotypes module
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+.. automodule:: happler.data.phenotypes
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+happler.tree.tree module
+~~~~~~~~~~~~~~~~~~~~~~~~
+
+.. automodule:: happler.tree.tree
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+happler.tree.treebuilder module
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+.. automodule:: happler.tree.treebuilder
    :members:
    :undoc-members:
    :show-inheritance:
