+++ docs/conf.py
@@ -45,6 +45,7 @@ exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
 
 # -- Extension configuration -------------------------------------------------
 autosummary_generate = True
+numpydoc_show_class_members = False 
 
 # -- Options for HTML output -------------------------------------------------
