+++ happler/data/__init__.py
@@ -1,0 +1,3 @@
+from .data import Data
+from .genotypes import Genotypes
+from .phenotypes import Phenotypes
