+++ happler/tree/tree.py
@@ -1,0 +1,18 @@
+import networkx as nx
+
+
+class Tree:
+    """
+    A tree where
+
+    1. nodes are variants (and also haplotypes)
+    2. each node can have at most two branches for each of its alleles
+
+    Attributes
+    ----------
+    graph: nx.DiGraph
+        The underlying directed acyclic graph representing the tree
+    """
+
+    def __init__(self):
+        self.graph = nx.DiGraph()
