+++ happler/tree/treebuilder.py
@@ -1,23 +1,5 @@
-import networkx as nx
-
-from .Data import Genotypes, Phenotypes
-
-
-class Tree:
-    """
-    A tree where
-
-    1. nodes are variants (and also haplotypes)
-    2. each node can have at most two branches for each of its alleles
-
-    Attributes
-    ----------
-    graph: nx.DiGraph
-        The underlying directed acyclic graph representing the tree
-    """
-
-    def __init__(self):
-        self.graph = nx.DiGraph()
+from .tree import Tree
+from ..data import Genotypes, Phenotypes
 
 
 class TreeBuilder:
