+++ tests/test_tree.py
@@ -2,8 +2,8 @@ import pytest
 import numpy as np
 from pathlib import Path
 
-from happler.Data import Genotypes, Phenotypes
-from happler.Tree import Tree, TreeBuilder
+from happler.data import Genotypes, Phenotypes
+from happler.tree import Tree, TreeBuilder
 
 
 DATADIR = Path(__file__).parent.joinpath("data")
@@ -14,6 +14,6 @@ def test_tree():
 
 
 def test_tree_builder():
-    gens = Genotypes.load(DATADIR.joinpath("simple.vcf"))
+    gens = Genotypes.load(DATADIR.joinpath("simple.vcf.gz"))
     phens = Phenotypes.load(DATADIR.joinpath("simple.tsv"))
     tree_builder = TreeBuilder(gens, phens)
